images/osm2pgsql/requirements.txt
psycopg2
mapcoloniesjsonlogger==0.0.14